# openapi/v1.yml
openapi: "3.0.2"
info:
  title: Events254
  description: Events254 internal API
  version: "1.0"
servers:
  - url: https://api.events254.co.ke
    description: Production servers
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ticket:
      type: object
      required:
        - id
        - event_id
        - price
        - currency
        - limit
        - type
      properties:
        id:
          type: integer
          format: int64
        event_id:
          type: integer
          format: int64
          description: The event that the ticket belogngs to
        price:
          type: number
          format: float
          description: The price of the current ticket
        type:
          type: string
          description: The type of ticket e.g Regular
          default: General admission
...
paths:
...
  /events/{eventId}/tickets/{ticketId}:
    get:
      description: Get a ticket by id
      operationId: Get ticket
      parameters:
        - in: path
          name: eventId
          required: true
          description: Event id
          schema:
            type: integer
            format: int64
        - in: path
          name: ticketId
          required: true
          description: Event id
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ticket"
        "404":
          description: Not found
        "401":
          description: Unauthenticated/Unauthorized
    put:
      description: update ticket information
      operationId: update ticket
      parameters:
        - in: path
          name: eventId
          required: true
          description: Event id
          schema:
            type: integer
            format: int64
        - in: path
          name: ticketId
          required: true
          description: Event id
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ticket"

        "401":
          description: Unauthenticated/Unauthorized
    delete:
      description: Delete a ticket
      operationId: delete ticket
      parameters:
        - in: path
          name: eventId
          required: true
          description: Event id
          schema:
            type: integer
            format: int64
        - in: path
          name: ticketId
          required: true
          description: Event id
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Deleted
          content:
            application/json:
              schema:
                type: string
        "404":
          description: ticket not found
        "401":
          description: Unauthenticated/Unauthorized
...